// features/web/grid/three-items.tsx
import type { Route } from "next";
import Link from "next/link";
import { GridTileImage } from "@/features/web/grid/tile";
import type { Product } from "@/shopify/types";

export function ThreeItemGridItem({
  item,
  size,
  priority,
}: {
  item: Product;
  size: "full" | "half";
  priority?: boolean;
}) {
  return (
    <div
      className={
        size === "full"
          ? "md:col-span-4 md:row-span-2"
          : "md:col-span-2 md:row-span-1"
      }
    >
      <Link
        className="relative block aspect-square h-full w-full"
        href={`/product/${item.handle}` as unknown as Route}
        prefetch={true}
      >
        <GridTileImage
          alt={item.title}
          fill
          label={{
            position: size === "full" ? "center" : "bottom",
            title: item.title as string,
            amount: item.priceRange.maxVariantPrice.amount,
            currencyCode: item.priceRange.maxVariantPrice.currencyCode,
          }}
          priority={priority}
          sizes={
            size === "full"
              ? "(min-width: 768px) 66vw, 100vw"
              : "(min-width: 768px) 33vw, 100vw"
          }
          src={item.featuredImage.url}
        />
      </Link>
    </div>
  );
}
